# Conflicts:
#	src/uet/oop/bomberman/BombermanGame.java
#	src/uet/oop/bomberman/entities/Bomber.java
#	src/uet/oop/bomberman/entities/Entity.java
#	src/uet/oop/bomberman/entities/bomb/Bomb.java

+++ src/uet/oop/bomberman/entities/AliveEntity.java
@@ -1,6 +1,6 @@
 package uet.oop.bomberman.entities;
 
 public interface AliveEntity {
-    void move(int x, int y);
+    void move(double x, double y);
     void setDead();
 }
